Fix MediaTags analyzer errors in OGG packet helpers
Move OggPacketInfo to its own file to satisfy MA0048 (file name matches type name) and update FindPacketIndex parameter type to List<OggPacketInfo> to satisfy CA1859.
